feat(ai): share the logo + nav header across all machine pages
Every /ai page now renders the BSMNT ASCII logo and the full section nav
(current section dimmed) instead of a per-page breadcrumb, so the machine
view is navigable from any page.
