+++ src/app/ai/faq/page.tsx
@@ -3,6 +3,7 @@ import { notFound } from "next/navigation"
 
 import { fetchFaqPage } from "@/app/(site)/(plain)/(content)/faq/sanity"
 import { Field, linkClass, Section } from "@/app/ai/components"
+import { MachineHeader } from "@/app/ai/machine-header"
 import { PageJsonLd } from "@/lib/structured-data/page-json-ld"
 
 export const metadata: Metadata = {
@@ -23,16 +24,7 @@ const MachineFaqPage = async () => {
       <PageJsonLd />
       <main className="mx-auto flex w-full max-w-2xl flex-col gap-8 px-4 pb-24 pt-12 text-f-p-mobile uppercase text-machine-base lg:text-f-p">
         <header className="flex flex-col gap-4">
-          <nav
-            aria-label="Machine index"
-            className="flex flex-wrap gap-x-4 gap-y-1"
-          >
-            <a href="/ai" className={linkClass}>
-              /ai
-            </a>
-            <span className="text-machine-dim">::</span>
-            <span className="text-machine-dim">faq</span>
-          </nav>
+          <MachineHeader current="/ai/faq" />
           <h1 className="text-machine-bright">
             basement.studio :: {faq.heading || "faq"}
           </h1>
