+++ src/app/ai/showcase/[slug]/page.tsx
@@ -8,6 +8,7 @@ import {
   getProjectData
 } from "@/app/(site)/(plain)/(content)/showcase/[slug]/sanity"
 import { Field, linkClass, Section } from "@/app/ai/components"
+import { MachineHeader } from "@/app/ai/machine-header"
 import { MachinePortableText } from "@/app/ai/machine-portable-text"
 import { extractPlainText } from "@/lib/structured-data/extract-text"
 import { PageJsonLd } from "@/lib/structured-data/page-json-ld"
@@ -50,20 +51,7 @@ const MachineProjectPage = async ({ params }: MachineProjectProps) => {
         {/* Header/meta stay uppercase like the /ai index; the project body keeps
           its authored casing for readability. */}
         <header className="flex flex-col gap-4 uppercase">
-          <nav
-            aria-label="Machine index"
-            className="flex flex-wrap gap-x-4 gap-y-1"
-          >
-            <a href="/ai" className={linkClass}>
-              /ai
-            </a>
-            <span className="text-machine-dim">::</span>
-            <a href="/ai/showcase" className={linkClass}>
-              showcase
-            </a>
-            <span className="text-machine-dim">::</span>
-            <span className="text-machine-dim">{project.slug}</span>
-          </nav>
+          <MachineHeader current="/ai/showcase" />
           <h1 className="text-machine-bright">{project.title}</h1>
           <dl className="flex flex-col gap-1">
             {project.client ? (
